[ll] Initial rework of command buffers
