+++ src/core/src/texture.rs
@@ -561,3 +561,18 @@ impl From<RenderDesc> for DepthStencilDesc {
         }
     }
 }
+
+///
+#[derive(Copy, Clone, Debug, Hash)]
+pub enum ImageLayout {
+    General,
+    ColorAttachmentOptimal,
+    DepthStencilAttachmentOptimal,
+    DepthStencilReadOnlyOptimal,
+    ShaderReadOnlyOptimal,
+    TransferSrcOptimal,
+    TransferDstOptimal,
+    Undefined,
+    Preinitialized,
+    Present,
+}
